Move selector to selector folder #3710
